Add MBDirectionOptions
